+++ creator/src/editcamera.rs
@@ -48,20 +48,27 @@ impl EditCamera {
         server_ctx: &mut ServerContext,
     ) {
         let mut view_switch = TheGroupButton::new(TheId::named("Editor View Switch"));
-        view_switch.add_text_status("2D".to_string(), fl!("status_camera_2d"));
-        let orbit_status = if cfg!(target_os = "macos") {
-            fl!("status_camera_orbit_macos")
-        } else {
-            fl!("status_camera_orbit_other")
-        };
-        let iso_status = if cfg!(target_os = "macos") {
-            fl!("status_camera_iso_macos")
-        } else {
-            fl!("status_camera_iso_other")
-        };
-        view_switch.add_text_status("Orbit".to_string(), orbit_status);
-        view_switch.add_text_status("Iso".to_string(), iso_status);
-        view_switch.add_text_status("FirstP".to_string(), fl!("status_camera_firstp"));
+        view_switch.add_text_status_icon(
+            String::new(),
+            fl!("tooltip_camera_2d"),
+            "square".to_string(),
+        );
+        view_switch.add_text_status_icon(
+            String::new(),
+            fl!("tooltip_camera_orbit"),
+            "perspective".to_string(),
+        );
+        view_switch.add_text_status_icon(
+            String::new(),
+            fl!("tooltip_camera_iso"),
+            "cube".to_string(),
+        );
+        view_switch.add_text_status_icon(
+            String::new(),
+            fl!("tooltip_camera_firstp"),
+            "camera".to_string(),
+        );
+        view_switch.set_item_width(30);
         view_switch.set_index(server_ctx.editor_view_mode.to_index());
         layout.add_widget(Box::new(view_switch));
         layout.set_reverse_index(Some(1));